Run tasks directly instead of using the sync method
Fixes #1291
