+++ core/src/main/java/com/github/games647/fastlogin/core/scheduler/AsyncScheduler.java
@@ -27,8 +27,6 @@ package com.github.games647.fastlogin.core.scheduler;
 
 import org.slf4j.Logger;
 
-import java.time.Duration;
-import java.util.concurrent.CompletableFuture;
 import java.util.concurrent.Executor;
 
 /**
@@ -46,30 +44,4 @@ public class AsyncScheduler extends AbstractAsyncScheduler {
             + "Upgrade Java to 21+ for improved performance");
     }
 
-    @Override
-    public CompletableFuture<Void> runAsync(Runnable task) {
-        return CompletableFuture.runAsync(() -> process(task), processingPool).exceptionally(error -> {
-            logger.warn("Error occurred on thread pool", error);
-            return null;
-        });
-    }
-
-    @Override
-    public CompletableFuture<Void> runAsyncDelayed(Runnable task, Duration delay) {
-        return CompletableFuture.runAsync(() -> {
-            currentlyRunning.incrementAndGet();
-            try {
-                Thread.sleep(delay.toMillis());
-                process(task);
-            } catch (InterruptedException interruptedException) {
-                // restore interrupt flag
-                Thread.currentThread().interrupt();
-            } finally {
-                currentlyRunning.getAndDecrement();
-            }
-        }, processingPool).exceptionally(error -> {
-            logger.warn("Error occurred on thread pool", error);
-            return null;
-        });
-    }
 }
